remove faucet add store code
