+++ packages/react-app/src/components/index.js
@@ -6,7 +6,6 @@ export { default as Blockie } from "./Blockie";
 export { default as BytesStringInput } from "./BytesStringInput";
 export { default as Contract } from "./Contract";
 export { default as EtherInput } from "./EtherInput";
-export { default as Faucet } from "./Faucet";
 export { default as GasGauge } from "./GasGauge";
 export { default as Header } from "./Header";
 export { default as Provider } from "./Provider";
