+++ packages/react-app/src/views/index.js
@@ -1,1 +1,2 @@
 export { default as Subgraph } from "./Subgraph";
+export { default as Store } from "./Store";
